move a function to a better shared location and add some docs
