+++ turbopack/crates/turbo-tasks-macros/src/func.rs
@@ -1195,3 +1195,48 @@ pub fn inline_inputs_identifier_filter(arg_ident: &Ident) -> bool {
     // filter out underscore-prefixed (unused) arguments, we don't need to cache these
     !arg_ident.to_string().starts_with('_')
 }
+
+/// Returns true if this attribute is a turbo_tasks attribute with the given name.
+fn is_attribute(attr: &Attribute, name: &str) -> bool {
+    let path = &attr.path();
+    if path.leading_colon.is_some() {
+        return false;
+    }
+    let mut iter = path.segments.iter();
+    match iter.next() {
+        Some(seg) if seg.arguments.is_empty() && seg.ident == "turbo_tasks" => match iter.next() {
+            Some(seg) if seg.arguments.is_empty() && seg.ident == name => iter.next().is_none(),
+            _ => false,
+        },
+        _ => false,
+    }
+}
+
+/// Parses a `turbo_tasks::function` attribute out of the given attributes and then returns the
+/// remaining attributes.
+pub fn split_function_attributes<'a>(
+    item: &'a impl Spanned,
+    attrs: &'a [Attribute],
+) -> (syn::Result<FunctionArguments>, Vec<&'a Attribute>) {
+    let (func_attrs_vec, attrs): (Vec<_>, Vec<_>) = attrs
+        .iter()
+        // TODO(alexkirsz) Replace this with function
+        .partition(|attr| is_attribute(attr, "function"));
+    let func_args = if let Some(func_attr) = func_attrs_vec.first() {
+        if func_attrs_vec.len() == 1 {
+            parse_with_optional_parens::<FunctionArguments>(func_attr)
+        } else {
+            Err(syn::Error::new(
+                // Report the error on the second annotation.
+                func_attrs_vec[1].span(),
+                "Only one #[turbo_tasks::function] attribute is allowed per method",
+            ))
+        }
+    } else {
+        Err(syn::Error::new(
+            item.span(),
+            "#[turbo_tasks::function] attribute missing",
+        ))
+    };
+    (func_args, attrs)
+}
